refactor: decompose large backend service files (#139)
Break down three oversized service files into focused, single-responsibility
modules while preserving the public API surface via facade pattern:

- drive.ts (921 lines) -> drive-token.ts (OAuth, tokens, connections) +
  drive-files.ts (folder/file CRUD) + drive.ts facade (147 lines)

- export.ts (621 lines) -> export-job.ts (PDF/EPUB generation) +
  export-delivery.ts (status, download, Drive upload) + export.ts facade (78 lines)

- source-material.ts (759 lines) -> source-types.ts (shared types) +
  source-drive.ts (Drive sources, folder expansion) +
  source-local.ts (local file upload/conversion) +
  source-material.ts facade (248 lines)

No route file or test file changes needed -- facades delegate to sub-services
and re-export all public types for backward compatibility.

Closes #139
